layout/style/test/chrome/test_bug1104916.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1104916
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1104916</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

/** Test for Bug 1104916 **/
SimpleTest.waitForExplicitFinish();
Components.utils.import("resource://gre/modules/Services.jsm");

function startTest() {
  var iframe = document.getElementById("subdoc");
  var subdoc = iframe.contentDocument;
  var style = subdoc.getElementById("style");
  var bodyComputedStyled = subdoc.defaultView.getComputedStyle(subdoc.body, "");

  function queryApplies(q) {
    style.setAttribute("media", q);
    return bodyComputedStyled.getPropertyValue("text-decoration") == "underline";
  }

  function shouldApply(q) {
    ok(queryApplies(q), q + " should apply");
  }

  function shouldNotApply(q) {
    ok(!queryApplies(q), q + " should not apply");
  }

  function waitForFullScreenChange() {
    return new Promise(function(resolve, reject) {
      document.addEventListener("mozfullscreenchange", function listener() {
        document.removeEventListener("mozfullscreenchange", listener);
        resolve();
      });
    });
  };

  shouldApply("all and (display-mode: browser)");
  shouldNotApply("all and (display-mode: fullscreen)");
  shouldNotApply("all and (display-mode: standalone)");
  shouldNotApply("all and (display-mode: minimal-ui)");

  var win = Services.wm.getMostRecentWindow("navigator:browser");
  // These view mode changes can be async
  var fullScreenChanged = waitForFullScreenChange();
  document.body.mozRequestFullScreen();
  fullScreenChanged.then(function() {
    ok(document.mozFullScreenElement, "window entered fullscreen");
    shouldApply("all and (display-mode: fullscreen)");
    shouldNotApply("all and (display-mode: browser)");
    fullScreenChanged = waitForFullScreenChange();
    document.mozCancelFullScreen();
    return fullScreenChanged;
  }).then(function() {
    ok(!document.mozFullScreenElement, "window exited fullscreen");
    shouldNotApply("all and (display-mode: fullscreen)");
    shouldApply("all and (display-mode: browser)");

    SimpleTest.finish();
  }).catch(function(reason) {
    ok(false, "promise rejection", reason)
    SimpleTest.finish();
  });
}
  </script>
</head>
<body onload="startTest();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1104916">Mozilla Bug 1104916</a>
<iframe id="subdoc" src="http://mochi.test:8888/tests/layout/style/test/media_queries_iframe.html"></iframe>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
